Pointing consumers to 'triplestore' for direct db-access
Consumers have default 'virtuoso' for direct database access.
In this app, 'triplestore' has been used. (In accordance with new project-template)
TODO: We need to bring this default up to date

Note: direct database access is only justified in very specific cases,
i.e. when we have performance problems.
This has a lot of rammifications so please do this wisely :-)