Use feature branch of resources and enable feature
This feature would allow strange resource types to be included and
returned from mu-cl-resources.  It may well be that the type of
resources the frontend would receive in this case is still not what is
to be expected but the chance of crashing resource itself through
inconsistent types would be lower at this point.